Fix removal of manifest file from sideload folder on Mac (#33)
On Mac, the manifest file is now renamed to ensure it is unique before being copied to the sideload folder. When removing the manifest file, ensure that we remove the manifest file even if not renamed, so we handle manifests added before the update to use a unique name.

Also added tests.